SQLite runtime (#14)
* SQLite runtime

* more unit tests

* add `createSchema`

* add / update documentation

* extract LogHandler; fix link in documentation; update sbt alias

* revert back some files

* fix compilation errors

* revert back documentation

* remove docs

* code review comments (wip)

* Remove redundant WorkflowId.scala and update SqliteRuntime

Removed the redundant WorkflowId.scala file. Updated the type
parameter in SqliteRuntime to be more specific by using String for
WorkflowId, ensuring consistency with typical workflow identifier
types.

* remove temporary examples until sqlruntime is ready

* Remove workflow_id from sqlite schema and storage

* Add PathCodec trait and apply it to SqliteRuntime

Added a new PathCodec trait to encode WorkflowId into a
string. Applied this trait as a context bound to the SqliteRuntime
class to handle the encoding of WorkflowId when creating
instances. This change ensures that WorkflowId is path-dependent and
extracted from the path as intended.

* Fix JDBC driver initialization in SqliteRuntime (wip)

* Add transaction mode to SQLite connection

* Add database path configuration to SqliteRuntime

* Remove PathCodec dependency and update db file handling

* unit tests (wip)

* Add beforeEach and afterEach hooks to SqliteSuite

* adapt storage

* fix compilation errors

* fix unit tests

* refactor sqlite runtime

* patch from Voytek